requirements.txt
absl-py==2.0.0
aiohttp==3.9.0
aiosignal==1.3.1
asgiref==3.7.2
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
ecdsa==0.18.0
firebase-admin==6.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
gast==0.5.4
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0
h5py==3.10.0
httplib2==0.22.0
idna==3.4
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
msgpack==1.0.7
multidict==6.0.4
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
openai==0.28.1
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
whitenoise==6.6.0
wikihowunofficialapi==0.1.1
wrapt==1.14.1
yarl==1.9.3
youtube-data-api==0.0.21
